+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 
 ### Fixed
+- Make the exception codes 32-bit safe (#100)
 - Always truncate the DB tables after an integration test (#86)
 
 ### Security